# perf optimizations