… .github/workflows/basic.yml
… .github/workflows/benchmark.yml
… .github/workflows/build_cpp.yml
… .github/workflows/build_python.yml
… CMakeLists.txt
… benchmarks/benchmark.cpp
… examples/novatel/converter_components/converter_components.cpp
… examples/novatel/converter_file_parser/converter_file_parser.cpp
… examples/novatel/converter_parser/converter_parser.cpp
… examples/novatel/json_parser/json_parser.cpp
… examples/novatel/range_decompressor/range_decompressor.cpp
… examples/novatel/rxconfig_handler/rxconfig_handler.cpp
… include/novatel_edie/common/circular_buffer.hpp
… include/novatel_edie/decoders/common/framer.hpp
… include/novatel_edie/decoders/common/message_decoder.hpp
… include/novatel_edie/decoders/oem/encoder.hpp
… include/novatel_edie/decoders/oem/filter.hpp
… include/novatel_edie/decoders/oem/header_decoder.hpp
… include/novatel_edie/decoders/oem/rxconfig/rxconfig_handler.hpp
… pyproject.toml
… python/bindings/decoder.cpp
… python/bindings/file_parser.cpp
… python/bindings/framer.cpp
… python/bindings/internal/decoder_tester.cpp
… python/bindings/message_database.cpp
… python/bindings/parser.cpp
… python/bindings/parser.hpp
… python/bindings/py_message_objects.cpp
… python/bindings/py_message_objects.hpp
… python/test/test_framer.py