I did some messing around with perf to see what the hotspots in our code are. The main changes are below.

- remove snprintf
- use simdjson over nlohmann
- avoid copying strings
- call the more efficient method of copying to a buffer when possible
- use move semantics to reduce ref counting for definitions
- use maps for O(1) lookups in enum definitions
- use from_chars instead of c functions

Creating less copies of the data and using doing stuff in place with string views is also a good idea for future improvement. If we could change the CRC polynomial, we could make binary encoding about 3x faster using hardware intrinsics.

Before:
![Screenshot from 2025-03-18 18-09-06](https://github.com/user-attachments/assets/f9156411-1cb0-4e57-89a5-4616516e19df)
After:
![Screenshot from 2025-04-11 17-41-42](https://github.com/user-attachments/assets/01bb703d-9d87-4fea-ab12-9b5eeeebf83b)