@rldoris 2025-03-13
Do you have any stats about the improvement in encoding speed that you could provide?