@jonathanmcdermid 2025-03-14
> Do you have any stats about the improvement in encoding speed that you could provide?

Before:
-------------------------------------------------------------------
Benchmark                         Time             CPU              Iterations
-------------------------------------------------------------------
EncodeAsciiLog                 2667 ns         2666 ns       261495       logs_per_second=375.044k/s
EncodeAbbrevAsciiLog    2367 ns         2366 ns       287237       logs_per_second=422.581k/s
EncodeJsonLog                 2689 ns         2689 ns       260123       logs_per_second=371.882k/s

After:
-------------------------------------------------------------------
Benchmark                         Time             CPU              Iterations
-------------------------------------------------------------------
EncodeAsciiLog                 1937 ns         1936 ns       362343        logs_per_second=516.535k/s
EncodeAbbrevAsciiLog     1618 ns        1617 ns       436032        logs_per_second=618.287k/s
EncodeJsonLog                  1941 ns         1941 ns       362965       logs_per_second=515.287k/s

25-35% drop in CPU time for ASCII encoding benchmarks, resulting in 35-50% increase in logs/second