@riley-kinahan 2025-04-10
Most of the changes look good to me. Main thing I'm not sure on is the switch to string views.

I don't understand how the lifetime of string data is managed now. It seems possible to me for a string views within a field container to outlive the actual data leading to a memory access error.

Also had a conversation with @rldoris this week about whether we can improve memory efficiency.  IIRC we have something like a 30x memory usage rate vs raw binary. One big culprit is that storing basic types like `uint64_t`s in a variant with `std::strings` and `std::vectors` leads to a lot of wasted bytes. Each `FieldContainer` is 64 bytes as of right now. Switching to `std::string_view`s gets us down to 56, but it might be worth considering using more primitive types for further savings.