@jonathanmcdermid 2025-04-11
>I don't understand how the lifetime of string data is managed now. It seems possible to me for a string_view within a FieldContainer to outlive the actual data leading to a memory access error.
@riley-kinahan Youve touched on an important limitation here, and its that our fieldcontainer values (not definitions) are only guaranteed to be valid while the message we are working with is still in the buffer and hasnt been overwritten. I think this is a limitation that we can work with, but it should be understood nonetheless.