@riley-kinahan 2025-04-11
> > I don't understand how the lifetime of string data is managed now. It seems possible to me for a string_view within a FieldContainer to outlive the actual data leading to a memory access error.
> 
> @riley-kinahan Youve touched on an important limitation here, and its that our fieldcontainer values (not definitions) are only guaranteed to be valid while the message we are working with is still in the buffer and hasnt been overwritten. I think this is a limitation that we can work with, but it should be understood nonetheless.
> 
> Additionally, I feel like this concept was already sort of baked in to the code. We preemptively call clear on the vector of fieldcontainers when we start decoding.

Ah ok, so my understanding is that during parsing messages will become invalid each time the framer buffer is overwritten. As a result it will be unsafe to access a message after the next one has been parsed.

There are a couple Python use cases which this would break, both of which I think are important to support:

1. Saving messages for delayed bulk processing:
```
bestpos_msgs = []
for msg in fp:
     if isinstance(msg, BESTPOS):
        bestpos_msgs.append(msg)
```

2. Multi-threaded processing:
```
while True:
    data = port.listen()
    parser.write(data)
    try:
        msg = parser.read()
        handler.put(msg)
    except ...
```

Do you have an estimate on how much efficiency is gained by avoiding string copying in particular?