@rldoris 2025-04-11
> > Also had a conversation with @rldoris this week about whether we can improve memory efficiency.  IIRC we have something like a 30x memory usage rate vs raw binary. One big culprit is that storing basic types like `uint64_t`s in a variant with `std::strings` and `std::vectors` leads to a lot of wasted bytes. Each `FieldContainer` is 64 bytes as of right now. Switching to `std::string_view`s gets us down to 56, but it might be worth considering using more primitive types for further savings.
> 
> An interesting problem. The best solution I managed to find was a custom variant implementation that automatically allocates all types above some size on the heap: https://cpp-rendering.io/small-object-optimization/ Definitely needs some benchmarking whether the extra indirection hurts the performance noticeably. I would assume it's a net benefit, though, since strings and vectors are not that often encountered in the messages.

@valgur The specific use case that developers are encountering is memory magnification when trying to store the data in Parquet/HD5 formats. It's more efficient to insert all the logs at once so they are writing code like this:

```messages = []
for message in parser.read():
    messages.append(message)
    
// Insert messages into Parquet file
```

and are seeing 32Gb systems run out of memory when processing 1.3Gb files. They've switched to processing the files in chunks, but the amplification is a bit too much so if we can minimize it I would like to try. Ultimately the best solution will be for us to document/provide examples for how to process large datasets because users will always generate large enough datasets that will overwhelm the memory of a PC.