@jonathanmcdermid 2025-04-11
@riley-kinahan performance gains from using string_view are significant, as string copying is expensive. I dont have the exact numbers available but it was one of the more significant changes made. I would say that if we need the behaviour your asking for, it would probably be best to do something like add a deep_copy() function for the fieldcontainers when you need them to outlive the lifespan of the message in the buffer, and only call it when needed. What do you think? @rldoris @valgur