@rldoris 2025-04-11
I was extremely happy to get rid of the `deepcopy()` functionality on the python side. Whenever I tried to introduce/explain the memory management to users on the python side I very quickly realized that was a big pain point for them. I would definitely cast my vote towards each object having safe/unique memory that doesn't require the user to manage. 

Another way to look at this too is that there aren't a ton of string fields in our most popular logs (VERSION is probably logged once so it doesn't count 😄) so having to copy string value probably isn't going to have a huge performance impact on average.